SerialPortTest: Implement main menu
In the course of improving the interface of the SerialPortTest application, I thought a main menu would be nice.
The menu was inspired by: https://wellsb.com/csharp/beginners/create-menu-csharp-console-application/

Option 3 (Send start to target value in time interval) is currently not implemented.

While at it also renamed the solution to c172hc-fsconnect.