Implement CalculateChecksum + StartToTargetValue
StartToTargetValue() asks the user for the start value, the end value
and the so called 'target time interval', the time  which should pass
between start and end value.

The message interval is (currently hardcoded) 50ms and the message value
is then sent to the c172hc-master in corresponding steps via the serial
interface.

Since the message is generated dynamically, the checksum must be
recalculated for each value. This is done in the CalculateChecksum
function.

This code was used for our first tests in the early morning of
2022-11-07. We noticed problems with the checksum calculation code
(either on the this (c172hc-fsconnect) or the c172hc-fsmaster side), but
ran out of time for further investigation.