Show serial ports setttings in menu + exit the application if no serial port is found
It makes no sense to start the application if no serial port is found. We then can assume that no c172hc-fsmaster is connected and exit the application.

We now default the serial port to the last one found (the highest COM port) because we assume that the last one is the one we got from c172hc-fsmaster and  then show the serial port settings in the main menu so we do not have to change the serial port settings if eveything is fine anyway.

Noticed that we queried the parity bit before the data bits when configuring the serial port. It it common practice to follow the order: COM, baud rate, data bits, parity, stop bits, handshake.

The class name PortChat was taken from SerialPort Class example (https://learn.microsoft.com/en-us/dotnet/api/system.io.ports.serialport). Time to move on...