Update CalculateChecksum function
Initially (in our meeting on 2022-07-06) we defined our c172hc-message-protocol
as follows:

- STX 0x02
- Gauge Code (1 Byte, i.e. 0x20 for Airspeed)
- Gauge Value (4 Byte)
- Checksum: 0xFF XOR GaugeCode -> XOR first Byte (MSB) -> XOR second byte -> XOR third byte -> XOR fourth byte
- ETX 0x03

The message-protocol and checksum implementation was tested against
c172hc-fsmaster which generated different checksums (see
jkirk/c172hc-fsconnect@05218eb).

It turned out that we had a misunderstanding about how to implement the
checksum and after a long debugging session we finally agreed on a calculation
based on the Intel Hex file checksum implementation:
https://tewarid.github.io/2013/08/01/calculate-intel-hex-file-checksum-using-windows-calculator.html

ch172hc-fsmaster was fixed partially (see: jkirk/c172hc-fsmaster@b8b101e): We
agreed to calculate the checksum using only the Gauge Code and Gauge Value (and
without STX (0x02)). Currently c172hc-fsmaster still uses STX to calculate the
checksum, this needs to be fixed.

Added a new Airspeed Demo Message function AirspeedDemoMessageCalculated()
where the given Gauge Type (0x20) and Gauge Value (0x1D 0x80) are used to
calculate the checksum and where the generated message is then sent to
c172hc-fsmaster.

The original function AirspeedDemoMessage() was renamed to
AirspeedDemoMessageHardcoded but basically does the same as before: It sends
the hard coded the Airspeed Demo Message to the serial port. The checksum
(0x43) was calculated by hand: NOT(0x20 + 0x1D + 0x80) + 1.

While at it also implemented a PrintMessage function which takes a byte array
message and prints the content of the message in hexadecimal format.