Continuously detect the presence of new or removed serial ports
It was quite annoying to restart the application when a c172hc-fsmaster board was disconnected or even connected.

Implemented a DetectSerialPorts() thread to monitor the presence of new or removed serial ports.
The "current serial port settings" in the main menu can be refreshed by reloading the main menu.
A new item "Reload Main Menu" has been added for this purpose.

For convenience, it is no longer necessary to press ENTER after selecting a menu item.