Add LatencyAnalysis (Send n messages @ m ms)
Ask the user for the gauge code, a start and a target value, a step value, a message interval and a no data timeout and perform a latency analysis when the start code is detected.

This means that a number of messages are sent to the c172hc-fsmaster within 10 seconds after the start code is received. We then compare the number of sent messages with the number of messages c172hc-fsmaster received.

If no data is received within the no data time interval (which defaults to 10 seconds) we timeout and start over.

Note, that we defined 0x55 as the start code but c172hc-fsmaster currently (bcca6e4) sends 0x55 via Serial.println() where it gets converted to the string "85\r\n". Therefor we need to trim the received line and compare it against "85".

Closes: jkirk/c172hc-fsconnect#1