Fix for loop overflow and underflow
When value hits 0, value -= step_value is still executed, causing value to underflow and get 2^32 -1.
As this is greater than target_value, the for loop never stops.

The same can happen with the value += step_value loop: if value hits 2^32-1 value + step_value is executed and value overflows.

So we check whether value is still greate or less than the start_value to prevent this from happening.

However, we should consider signed values.