Improve Latency Analysis
Show how many messages have been sent in how many ms.
We now allow the user to change start_value for the next message,
but we have changed the default target_value to start_value + 9 (so that 10 messages are sent by default).
The other values default to the values from the previous run.

This makes it easier to repeat similar messages.