Use RequestDataOnSimObject instead of RequestDataOnSimObjectType
The SimConnect_RequestDataOnSimObject function is used to request when the SimConnect client is to receive data values for a specific object.
The SimConnect_RequestDataOnSimObjectType function is used to retrieve information about simulation objects of a given type that are within a specified radius of the user's aircraft.

So it makes more sense to use RequestDataOnSimObject as we only need the data values from the user's aircraft and not from any other object (within any specified radius).

The unit of altitude + airspeed is also displayed.