Request Cessna Six Pack values
We want to receive the values of the so-called Cessna Six Pack, the primary flight instruments.
These are the following:

- Airspeed Indicator
- Attitude Indicator (pitch + bank)
- Altimeter
- Vertical Speed Indicator
- Heading Indicator
- Turn Coordinator

We do not need latitude and longitude, so remove them.
We also do not want to output all previous values, only the most recent.
We also allow the user to stop requesting new data.