Request MakeFaire specific values
For the MakeFaire we decided to show some values that are easier to display.
These are the following:
- Airspeed
- Altimeter
- Vertical speed
- RPM

So we implemented a second button that only requests for these MakerFaire specific values.
We use integer precision instead of float because analogue instruments do not have such a high resolution anyway.
Only one request can be active at a given time.