On SendMessage try to open the serial port
If the serial port was disconnected or the configuration changed, the serial connection was not re-established.

Previously, we opened the serial connection when the port was newly detected or when the configuration changed. But if the serial port was lost and we got the same port again, we did not open it again.
We now open the serial connection (if it was not already open) every time we send a message.

Also, disable the SendSingleMessageBlock before we start the fsmaster class,
because there was a race condition where the serial port was detected
and the SendSingleMessageBlock button was enabled, but we disabled it immediately afterwards.